Fixed a bug in the calculation of the size of the label array. The bug caused programs like     L F = 1;     label X1;     label X2;     label X3;     label X4;     label X5;     label X6;     label X7;     label X8;     .sort;     .end to crash on 32bit systems.

+++ sources/names.c
@@ -2078,7 +2078,7 @@ int GetLabel(UBYTE *name)
 			MesPrint("&More than 32765 labels in one module. Please simplify.");
 			Terminate(-1);
 		}
-		NewLabelNames = (UBYTE **)Malloc1((sizeof(UBYTE *)+sizeof(WORD))
+		NewLabelNames = (UBYTE **)Malloc1((sizeof(UBYTE *)+sizeof(int))
 				*newnum,"Labels");
 		NewLabel = (int *)(NewLabelNames+newnum);
 		for ( i = 0; i< AC.MaxLabels; i++ ) {
